Add new files including vim packages
This can be thought of as kind of a second initial commit, as the first
one was not very substantive. There is a lot of work left to be done to
clean these up and decide what I really like. I expect to change this
quite often for a long time.